+++ www/migration.php
@@ -1,20 +1,21 @@
 #!/usr/bin/php
 <?php
-/** \file
- *  Springy.
+
+/**
+ * Command line migration script launcher.
  *
- *	\brief     This is a database migration script
- *  \copyright ₢ 2007-2016 Fernando Val
- *  \author    Fernando Val - fernando.val@gmail.com
- *  \version   0.4.0.4
- *  \ingroup   framework
+ * @copyright 2007 Fernando Val
+ * @author    Fernando Val <fernando.val@gmail.com>
+ *
+ * @version 0.4.1
  */
-if (!file_exists(__DIR__ . '/sysconf.php')) {
+
+if (!file_exists('sysconf.php')) {
     echo 'Internal System Error on Startup.',"\n";
     echo 'Required file "sysconf.php" missing.',"\n";
     exit(999);
 }
-if (!file_exists(__DIR__ . '/_Main.php')) {
+if (!file_exists('_Main.php')) {
     echo 'Internal System Error on Startup.',"\n";
     echo 'Required file "_Main.php" missing.',"\n";
     exit(999);